refactor: relocate cmds implementation
