+++ lib/src/commands/x/rm/rm.dart
