+++ lib/src/commands/x/x.dart
@@ -1,5 +1,5 @@
 import 'package:args/command_runner.dart';
-import 'package:cov_utils/src/x/rm/rm.dart';
+import 'package:cov_utils/src/commands/x/rm/rm.dart';
 
 /// A command invocation function that provides generic functionalities.
 Future<void> x(List<String> args) async {
